include,client,mds: relocate filepath header
This is an internal data structure that is not used outside of Ceph.
