+++ src/crimson/CMakeLists.txt
@@ -102,7 +102,7 @@ add_library(crimson-common STATIC
   ${PROJECT_SOURCE_DIR}/src/crush/CrushTester.cc
   ${PROJECT_SOURCE_DIR}/src/global/global_context.cc
   ${PROJECT_SOURCE_DIR}/src/global/pidfile.cc
-  ${PROJECT_SOURCE_DIR}/src/include/filepath.cc
+  ${PROJECT_SOURCE_DIR}/src/common/filepath.cc
   ${PROJECT_SOURCE_DIR}/src/include/object.cc
   ${PROJECT_SOURCE_DIR}/src/include/types.cc
   ${PROJECT_SOURCE_DIR}/src/include/utime.cc
