+++ src/mds/CInode.h
@@ -31,7 +31,7 @@
 #include "include/Context.h" // for C_GatherBuilder
 #include "include/counter.h"
 #include "include/elist.h"
-#include "include/filepath.h"
+#include "common/filepath.h"
 #include "include/types.h"
 #include "include/compact_set.h"
