+++ src/mds/MDCache.h
@@ -26,7 +26,7 @@
 #include "common/admin_finisher.h"
 #include "include/common_fwd.h"
 #include "include/types.h"
-#include "include/filepath.h"
+#include "common/filepath.h"
 #include "include/elist.h"
 #include "include/rados/rados_types.hpp"
