+++ src/mds/Migrator.cc
@@ -28,7 +28,7 @@
 #include "MDSMap.h"
 #include "Mutation.h"
 
-#include "include/filepath.h"
+#include "common/filepath.h"
 #include "common/debug.h"
 #include "common/likely.h"
