+++ src/messages/MDiscoverReply.h
@@ -17,7 +17,7 @@
 #ifndef CEPH_MDISCOVERREPLY_H
 #define CEPH_MDISCOVERREPLY_H
 
-#include "include/filepath.h"
+#include "common/filepath.h"
 #include "messages/MMDSOp.h"
 
 #include <string>
