+++ src/test/libcephfs/test.cc
@@ -14,7 +14,7 @@
  */
 
 #include "include/compat.h"
-#include "include/filepath.h"
+#include "common/filepath.h"
 #include "gtest/gtest.h"
 #include "include/cephfs/libcephfs.h"
 #include "mds/mdstypes.h"
