+++ src/test/old/testfilepath.cc
@@ -1,5 +1,5 @@
 
-#include "include/filepath.h"
+#include "common/filepath.h"
 #include <iostream>
 using namespace std;
